requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
requests==2.31.0
redis==5.0.1
aiohttp==3.9.1
python-jose==3.3.0
websockets==12.0
numpy==1.26.2
pandas==2.1.3
